+++ packages/string-collapse-white-space/changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - 2017-09-16
+### Added
+- ✨ `opts.trimLines` - activates trim per-line basis
+- ✨ `opts.trimnbsp` - non-breaking spaces are trimmed too
+- ✨ switched to ESLint on `airbnb-base` config, with 3 exceptions: 1. no semicolons; 2. plus-plus allowed in loops; 3. no short identifiers (with some exceptions)
+
 ## [2.1.0] - 2017-09-03
 ### Added
 - ✨ Correctly treats non-breaking spaces - they are not considered _collapsable_ or _trimmable_ now.
@@ -16,5 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### New
 - ✨ First public release
 
+[2.2.0]: https://github.com/codsen/string-collapse-white-space/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/codsen/string-collapse-white-space/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/codsen/string-collapse-white-space/compare/v1.0.0...v2.0.0